requirements.txt
Django==2.1.14
djangorestframework==3.9.1
Pillow==6.2.1
psycopg2==2.8.4
requests==2.22.0
urllib3==1.25.7